simple_calculator
This is a basic and simple calculator designed through the use of Python programming language. In this program I have used conditional statements. A user can modify this program according to his own needs by using functions, importing the math library and can also increase the functionality of this program but adding more features to it!